Fixed STSMiner wrong connection and wrong kind of statement bugs.
STSMiner was using the wrong connection to the database. The algorithm has been fixed so to use the connection defined in MiningAlgorithmBase instead of using the one defined in AlgorithmsOptions. The connection in AlgorithmsOptions has been removed since it is not used.

STSMine was using executeQuery API to execute insert/create/delete statements. These have been fixed to use the correct API.